fix: More defensive code when trying to drop tables

+++ migration/2_nag_upgrade_autoincrement.php
@@ -20,9 +20,8 @@ class NagUpgradeAutoIncrement extends Horde_Db_Migration_Base
     public function up()
     {
         $this->changeColumn('nag_shares', 'share_id', 'autoincrementKey');
-        try {
+        if (in_array('nag_shares_seq', $this->tables())) {
             $this->dropTable('nag_shares_seq');
-        } catch (Horde_Db_Exception $e) {
         }
     }
